+++ fastapi_app/src/domain/location/use_cases/delete_location.py
@@ -1,7 +1,10 @@
 from infrastructure.sqlite.database import database
 from infrastructure.sqlite.repositories.locations import LocationRepository
-from core.exceptions.domain_exceptions import LocationNotFoundByIdException
+from core.exceptions.domain_exceptions import LocationNotFoundByIdException, ForbiddenException
 from core.exceptions.database_exceptions import LocationNotFoundException
+import logging
+
+logger = logging.getLogger(__name__)
 
 
 class DeleteLocationUseCase:
@@ -9,7 +12,11 @@ class DeleteLocationUseCase:
         self._database = database
         self._repo = LocationRepository()
 
-    async def execute(self, location_id: int) -> bool:
+    async def execute(self, location_id: int, is_superuser: bool = False) -> bool:
+        if not is_superuser:
+            error = ForbiddenException()
+            logger.error("Только администратор может удалять локации")
+            raise error
         with self._database.session() as session:
             try:
                 self._repo.delete(session=session, id=location_id)
