+++ fastapi_app/src/domain/post/use_cases/delete_post.py
@@ -1,7 +1,10 @@
 from infrastructure.sqlite.database import database
 from infrastructure.sqlite.repositories.posts import PostRepository
-from core.exceptions.domain_exceptions import PostNotFoundByIdException
+from core.exceptions.domain_exceptions import PostNotFoundByIdException, ForbiddenException
 from core.exceptions.database_exceptions import PostNotFoundException
+import logging
+
+logger = logging.getLogger(__name__)
 
 
 class DeletePostUseCase:
@@ -9,7 +12,15 @@ class DeletePostUseCase:
         self._database = database
         self._repo = PostRepository()
 
-    async def execute(self, post_id: int) -> bool:
+    async def execute(self,
+                      post_id: int,
+                      author_id: int,
+                      current_user_id: int,
+                      is_superuser: bool = False) -> bool:
+        if author_id != current_user_id and not is_superuser:
+            error = ForbiddenException()
+            logger.error("Только автор поста и администратор может удалить его")
+            raise error
         try:
             with self._database.session() as session:
                 self._repo.delete(session=session, id=post_id)
